Move mock test to __mock__
