+++ src/__mocks__/leaflet.js
