+++ src/components/review/review.test.js
@@ -1,5 +1,5 @@
 import {Review} from './review';
-import {TestMock} from '../../test-mock/test-mock';
+import {TestMock} from '../../__mocks__/mocks';
 
 describe(`<Review/>: `, () => {
   it(`should render correctly`, () => {
